Creates PDF/ZIP only if user ticks on it.
merge_badges.py file is taken from https://github.com/fossasia/badgeyay/pull/71 and further tweaking is done on it.
index.html is changed to be updated with the new changes. and main.py is also changed. style.css is also changed with accordance to it.

resolved conflicts as well

Closes https://github.com/fossasia/badgeyay/issues/66