Generate Badges for Manual Data with pre-selected images
Changed an elif to if so that if a predefined image is used the the CSV is also checked for.
The error was caused because of the fact that elif was not executed if the if of image was encountered.
I've managed it and fixed the bug. There were several other bugs as well that I've fixed so as to implement proper functionality.
All the changes can be found in the commit section on thie PR

Fixes : https://github.com/fossasia/badgeyay/issues/314