app/merge_badges.py: Replaces pdftk with pypdf (#410)
pdftk which is a non-python library was used to merge pdf files. So
every time to merge pdf files it has to be run as a separate process
via subprocess as it is not a python native library. Replaces it with
pypdf which is a similar library but in python that can be used to merge
pdf files.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>