Migrates to REST API
Converts the existing Flask app which was coupled into a back-end which
acts as a REpresentational State Transfer based API. Provides all
functionalities of the app to be provided over the new API.

Adds and modifies related documentation.

The tests have been commented out from running in the CI platforms like
.travis.yml as they are specific to the coupled system and not with the
back-end alone.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>